Revert addition of `__lt__` to `Record_SRV`.
